feat: audio added

+++ include/sa_algo.h
@@ -26,7 +26,8 @@
 
 
 
-#define SA_Normalize_I(val, max)     ((val)/(max))
+
+#define SA_Normalize_I(val, max) ((val) / (float)(max))
